+++ json_annotation/lib/json_annotation.dart
@@ -12,6 +12,7 @@ library json_annotation;
 
 export 'src/allowed_keys_helpers.dart';
 export 'src/checked_helpers.dart';
+export 'src/enum_helpers.dart';
 export 'src/json_converter.dart';
 export 'src/json_enum.dart';
 export 'src/json_key.dart';