+++ json_serializable/lib/src/type_helpers/enum_helper.dart
@@ -45,14 +45,11 @@ class EnumHelper extends TypeHelper<TypeHelperContextWithConfig> {
       return null;
     }
 
-    context.addMember(enumDecodeHelper);
-
     String functionName;
     if (targetType.isNullableType || defaultProvided) {
-      functionName = r'_$enumDecodeNullable';
-      context.addMember(_enumDecodeHelperNullable);
+      functionName = r'$enumDecodeNullable';
     } else {
-      functionName = r'_$enumDecode';
+      functionName = r'$enumDecode';
     }
 
     context.addMember(memberContent);
@@ -86,42 +83,3 @@ String? enumValueMapFromType(DartType targetType) {
 
   return 'const ${_constMapName(targetType)} = {\n$items\n};';
 }
-
-const enumDecodeHelper = r'''
-K _$enumDecode<K, V>(
-  Map<K, V> enumValues,
-  Object? source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    throw ArgumentError(
-      'A value must be provided. Supported values: '
-      '${enumValues.values.join(', ')}',
-    );
-  }
-
-  return enumValues.entries.singleWhere(
-    (e) => e.value == source,
-    orElse: () {
-      if (unknownValue == null) {
-        throw ArgumentError(
-          '`$source` is not one of the supported values: '
-          '${enumValues.values.join(', ')}',
-        );
-      }
-      return MapEntry(unknownValue, enumValues.values.first);
-    },
-  ).key;
-}''';
-
-const _enumDecodeHelperNullable = r'''
-K? _$enumDecodeNullable<K, V>(
-  Map<K, V> enumValues,
-  dynamic source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    return null;
-  }
-  return _$enumDecode<K, V>(enumValues, source, unknownValue: unknownValue);
-}''';