+++ json_serializable/test/default_value/default_value.g_any_map__checked.g.dart
@@ -50,7 +50,7 @@ DefaultValue _$DefaultValueFromJson(Map json) => $checkedCreate(
                     'root': ['child']
                   }),
           $checkedConvert('fieldEnum',
-              (v) => _$enumDecodeNullable(_$GreekEnumMap, v) ?? Greek.beta),
+              (v) => $enumDecodeNullable(_$GreekEnumMap, v) ?? Greek.beta),
           constClass: $checkedConvert(
               'constClass',
               (v) => v == null
@@ -90,43 +90,6 @@ Map<String, dynamic> _$DefaultValueToJson(DefaultValue instance) =>
           const ConstClassConverter().toJson(instance.valueFromConverter),
       'valueFromFunction': constClassToJson(instance.valueFromFunction),
     };
-
-K _$enumDecode<K, V>(
-  Map<K, V> enumValues,
-  Object? source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    throw ArgumentError(
-      'A value must be provided. Supported values: '
-      '${enumValues.values.join(', ')}',
-    );
-  }
-
-  return enumValues.entries.singleWhere(
-    (e) => e.value == source,
-    orElse: () {
-      if (unknownValue == null) {
-        throw ArgumentError(
-          '`$source` is not one of the supported values: '
-          '${enumValues.values.join(', ')}',
-        );
-      }
-      return MapEntry(unknownValue, enumValues.values.first);
-    },
-  ).key;
-}
-
-K? _$enumDecodeNullable<K, V>(
-  Map<K, V> enumValues,
-  dynamic source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    return null;
-  }
-  return _$enumDecode<K, V>(enumValues, source, unknownValue: unknownValue);
-}
 
 const _$GreekEnumMap = {
   Greek.alpha: 'alpha',