+++ json_serializable/test/supported_types/input.type_enumtype.g.dart
@@ -9,8 +9,8 @@ part of 'input.type_enumtype.dart';
 // **************************************************************************
 
 SimpleClass _$SimpleClassFromJson(Map<String, dynamic> json) => SimpleClass(
-      _$enumDecode(_$EnumTypeEnumMap, json['value']),
-      _$enumDecodeNullable(_$EnumTypeEnumMap, json['withDefault']) ??
+      $enumDecode(_$EnumTypeEnumMap, json['value']),
+      $enumDecodeNullable(_$EnumTypeEnumMap, json['withDefault']) ??
           EnumType.alpha,
     );
 
@@ -20,32 +20,6 @@ Map<String, dynamic> _$SimpleClassToJson(SimpleClass instance) =>
       'withDefault': _$EnumTypeEnumMap[instance.withDefault],
     };
 
-K _$enumDecode<K, V>(
-  Map<K, V> enumValues,
-  Object? source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    throw ArgumentError(
-      'A value must be provided. Supported values: '
-      '${enumValues.values.join(', ')}',
-    );
-  }
-
-  return enumValues.entries.singleWhere(
-    (e) => e.value == source,
-    orElse: () {
-      if (unknownValue == null) {
-        throw ArgumentError(
-          '`$source` is not one of the supported values: '
-          '${enumValues.values.join(', ')}',
-        );
-      }
-      return MapEntry(unknownValue, enumValues.values.first);
-    },
-  ).key;
-}
-
 const _$EnumTypeEnumMap = {
   EnumType.alpha: 'alpha',
   EnumType.beta: 'beta',
@@ -53,21 +27,10 @@ const _$EnumTypeEnumMap = {
   EnumType.delta: 'delta',
 };
 
-K? _$enumDecodeNullable<K, V>(
-  Map<K, V> enumValues,
-  dynamic source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    return null;
-  }
-  return _$enumDecode<K, V>(enumValues, source, unknownValue: unknownValue);
-}
-
 SimpleClassNullable _$SimpleClassNullableFromJson(Map<String, dynamic> json) =>
     SimpleClassNullable(
-      _$enumDecodeNullable(_$EnumTypeEnumMap, json['value']),
-      _$enumDecodeNullable(_$EnumTypeEnumMap, json['withDefault']) ??
+      $enumDecodeNullable(_$EnumTypeEnumMap, json['value']),
+      $enumDecodeNullable(_$EnumTypeEnumMap, json['withDefault']) ??
           EnumType.alpha,
     );