+++ json_annotation/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added `JsonSerializabel.constructor` field to allow specifying an alternative
   constructor to invoke when creating a `fromJson` helper.
 - Added `JsonEnum` for annotating `enum` types.  
+- Require Dart SDK `>=2.14.0`.
 
 ## 4.1.0