+++ json_annotation/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 - Added `JsonSerializabel.constructor` field to allow specifying an alternative
   constructor to invoke when creating a `fromJson` helper.
-- Added `JsonEnum` for annotating `enum` types.  
+- Added `JsonEnum` for annotating `enum` types.
+- Added `$enumDecodeNullable` and `$enumDecode` helpers to minimize generated
+  code.
 - Require Dart SDK `>=2.14.0`.
 
 ## 4.1.0